+++ Runtime/CoreSystemSettings.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName ="CoreSystemSettings", menuName ="MyPackage/Core System Settings")]
+public class CoreSystemSettings : ScriptableObject
+{
+    public GameObject coreSystemPrefab;
+}